refactor(front): Renamed `Snackbar` to `AppSnackbar`
